Add initial draft of absorbing non-monotonic functions into thresholds
This generalizes the _AbsorbFunctionIntoComparison rule to allow for
increasing and decreasing branches of inverse functions and adds a
threshold fusion rule to eventually combine the resulting segments.

This still needs to be extended to support n-ary functions and node
attributes to also apply this to functions such as Swish or Gelu.

Currently the new functionality can be demonstrated with Silu and x^2,
and arbitrary combinations of these and already supported functions.